requirements.txt
aiogram==3.4.1
httpx>=0.24.0
python-dotenv>=1.0.0
